Commit apres test config et changement idUser to int

+++ src/main/java/tn/isetsf/presence/Repository/UserRepo.java
@@ -6,7 +6,7 @@ import tn.isetsf.presence.Entity.Users;
 
 import java.util.Optional;
 
-public interface UserRepo extends JpaRepository<Users, Long> {
+public interface UserRepo extends JpaRepository<Users, Integer> {
     Optional<Users> findByLoginAndPassword(String login, String password);
 
 }
